src/requirements.txt
# Python requirements file for RaspberryPi/Server

pyserial
python-dotenv
adafruit-io
redis
RPi.GPIO
